requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
opencv-python>=4.8
albumentations==1.3.1
pillow>=9.5
matplotlib>=3.7
seaborn>=0.12
tensorboard>=2.14
simpleitk>=2.2
monai>=1.3
captum>=0.7
pyyaml>=6.0
rich>=13.5
tqdm>=4.66
